renamed and tested warnings
