+++ algos/starter-algo/gamelib/__init__.py
@@ -4,10 +4,10 @@ The gamelib package contains modules that assist in algo creation
 
 from .algocore import AlgoCore
 from .util import debug_write
-from .game import GameState
+from .game_state import GameState
 from .unit import GameUnit
 from .game_map import GameMap
-from .advanced import AdvancedGameState
+from .advanced_game_state import AdvancedGameState
 
-__all__ = ["advanced", "algocore", "game", "game_map", "navigation", "unit", "util"]
+__all__ = ["advanced_game_state", "algocore", "game_state", "game_map", "navigation", "unit", "util"]
